Big refactor finished
Now uses a model where Race, Stage and Image objects wrap underlying
data from cyclingstage.com and procyclingstats.com

External resources are downloaded with get_resource() which exposes the
underlying model.

Data is processed for most available attrs on the fly.

The idea is that failures are just not downloaded and this is apparent
on the filesystem.

Seems to work fully.